Updating the step-by-step tutorial
Updating includes doing the tutorial on myself step by step.
I created a few new screenshots (mainly related to layout of the GUI)
Also the starting point opensesame files are new (I'm not sure whether
this is really necessary)
and of course the actual text.
NOTE: Because of this issue:
http://forum.cogsci.nl/index.php?p=/discussion/1039/open-problems-with-running-winpython-2-8-1
I was not able to test whether the experiment really works in the end.
Sorry for that.